perf(nodes): improve performance for large node lists and reduce re-renders
