+++ app/dashboard/package.json
@@ -35,7 +35,7 @@
     "react-hook-form": "^7.39.1",
     "react-i18next": "^12.2.0",
     "react-json-editor-ajrm": "^2.5.14",
-    "react-loading-skeleton": "^3.1.0",
+    "react-loading-skeleton": "^3.5.0",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.4.3",
     "react-slick": "^0.29.0",
